solved multiple chat creation issue in group chat

+++ frontend/my-app/src/components/ChatWindow.js
@@ -4,32 +4,26 @@ import axios from 'axios';
 import logo from '../assets/logo.png';
 import { UserContext } from '../context/UserContext';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faCheck, faL, faTicket, faTrashAlt } from '@fortawesome/free-solid-svg-icons';
-import io from 'socket.io-client'
+import { faCheck, faTrashAlt } from '@fortawesome/free-solid-svg-icons';
+import io from 'socket.io-client';
 
-
-const ENDPOINT = 'http://localhost:4000'
+const ENDPOINT = 'http://localhost:4000';
 let socket;
 
-function ChatWindow({ selectedChat, setSelectedChat, setUsersListUpdate}) {
-
+function ChatWindow({ selectedChat, setSelectedChat, setUsersListUpdate }) {
+    const groupIconUrl = 'https://static.vecteezy.com/system/resources/previews/026/019/617/non_2x/group-profile-avatar-icon-default-social-media-forum-profile-photo-vector.jpg';
 
     const { userDetails } = useContext(UserContext);
     const [otherUserDetails, setOtherUserDetails] = useState({});
-    const [chatDetails, setChatDetails] = useState({});
-
-
+    const [chatDetails, setChatDetails] = useState(null);
     const [messages, setMessages] = useState([]);
     const [newMessage, setNewMessage] = useState('');
-
-
-    const [socketConnection, setSocketConnection] = useState(false)
-
+    const [socketConnection, setSocketConnection] = useState(false);
     const selectedChatCompare = useRef(null);
-    const messagesEndRef = useRef(null)
-    const [loading, setLoading] = useState(false)
-    const [typing, setTyping] = useState(false)
-    const [isTyping, setIsTyping] = useState(false)
+    const messagesEndRef = useRef(null);
+    const [loading, setLoading] = useState(false);
+    const [typing, setTyping] = useState(false);
+    const [isTyping, setIsTyping] = useState(false);
 
     const config = useMemo(() => ({
         headers: {
@@ -40,52 +34,49 @@ function ChatWindow({ selectedChat, setSelectedChat, setUsersListUpdate}) {
 
     useEffect(() => {
         if (selectedChat) {
-            fetchChat()
+            fetchChat();
         }
-    }, [selectedChat])
-
+    }, [selectedChat]);
 
     const scrollToBottom = () => {
-        messagesEndRef.current?.scrollIntoView({ behavior: "smooth" })
-    }
+        messagesEndRef.current?.scrollIntoView({ behavior: "smooth" });
+    };
 
     useEffect(() => {
-        scrollToBottom()
+        scrollToBottom();
     }, [messages]);
 
     useEffect(() => {
-        socket = io(ENDPOINT)
-        socket.emit('setupSocket', userDetails)
+        socket = io(ENDPOINT);
+        socket.emit('setupSocket', userDetails);
         socket.on('connected', () => {
-            console.log("socket connected")
-            setSocketConnection(true)
-        })
+            console.log("socket connected");
+            setSocketConnection(true);
+        });
 
-        socket.on("typing", () => { setIsTyping(true) })
-        socket.on("stop typing", () => { setIsTyping(false) });
+        socket.on("typing", () => { setIsTyping(true); });
+        socket.on("stop typing", () => { setIsTyping(false); });
 
         socket.on('messageReceived', (newMessageReceived) => {
-            var newMessageId = newMessageReceived.chat._id
-            console.log(selectedChatCompare.current === newMessageId)
-            if (selectedChatCompare && selectedChatCompare.current === newMessageId) {
-                console.log("message received")
+            var newMessageId = newMessageReceived.chat._id;
+            if (selectedChatCompare.current === newMessageId) {
+                console.log("message received");
                 setMessages((prevMessages) => [...prevMessages, newMessageReceived]);
             }
-            else {
-
-            }
         });
+
         return () => {
             socket.disconnect();
         };
-    }, [userDetails])
+    }, [userDetails]);
 
     function handleDelete() {
+        if (!chatDetails || !chatDetails._id) return;
         axios.delete(`http://localhost:4000/api/chat/${chatDetails._id}`, config)
             .then((response) => {
                 if (response.status === 200) {
                     setSelectedChat(null);
-                    setUsersListUpdate(prev=> !prev)
+                    setUsersListUpdate(prev => !prev);
                 }
             })
             .catch((err) => {
@@ -94,13 +85,13 @@ function ChatWindow({ selectedChat, setSelectedChat, setUsersListUpdate}) {
     }
 
     function handleSendMessage() {
-        if (newMessage.trim() === '') return;
-        socket.emit('stop typing', chatDetails._id)
+        if (newMessage.trim() === '' || !chatDetails || !chatDetails._id) return;
+        socket.emit('stop typing', chatDetails._id);
         axios.post('http://localhost:4000/api/message', { content: newMessage, chatId: chatDetails._id }, config)
             .then((response) => {
                 setMessages([...messages, response.data]);
                 setNewMessage('');
-                socket.emit('newMessage', response.data)
+                socket.emit('newMessage', response.data);
             })
             .catch((err) => {
                 console.log(err);
@@ -108,95 +99,88 @@ function ChatWindow({ selectedChat, setSelectedChat, setUsersListUpdate}) {
     }
 
     function handleInputChange(e) {
-        setNewMessage(e.target.value)
+        setNewMessage(e.target.value);
 
-        if (!socketConnection) return
+        if (!socketConnection || !chatDetails || !chatDetails._id) return;
 
         if (!typing) {
-            setTyping(true)
-            socket.emit('typing', chatDetails._id)
+            setTyping(true);
+            socket.emit('typing', chatDetails._id);
         }
 
         let lastTypingTime = new Date().getTime();
         setTimeout(() => {
-            var currentTime = new Date().getTime()
+            var currentTime = new Date().getTime();
             if (currentTime - lastTypingTime >= 2000 && typing) {
-                socket.emit('stop typing', chatDetails._id)
-                setTyping(false)
+                socket.emit('stop typing', chatDetails._id);
+                setTyping(false);
             }
-        }, 2000)
+        }, 2000);
     }
 
     function fetchChat() {
         axios.get(`http://localhost:4000/api/chat/${selectedChat}`, config)
             .then((response) => {
                 const chatData = response.data;
-                console.log(chatData)
-                setChatDetails(chatData)
+                console.log(response.data)
+                setChatDetails(chatData);
 
-                // //Other User Details
-                const otherUser = chatData.users.find((user) => user._id !== userDetails._id);
-                setOtherUserDetails(otherUser)
-                selectedChatCompare.current = chatData._id
+                if (chatData.isGroupChat) {
+                    const otherGroupUsers = chatData.users.filter((user) => user._id !== userDetails._id);
+                    setOtherUserDetails([otherGroupUsers]);
+                } else {
+                    const otherUser = chatData.users.find((user) => user._id !== userDetails._id);
+                    setOtherUserDetails(otherUser);
+                }
+
+                selectedChatCompare.current = chatData._id;
             })
             .catch((err) => {
-                console.log(err)
-            })
+                console.log(err);
+            });
     }
 
-    // get chat details GET - api/message/:id
     useEffect(() => {
-        setLoading(true)
-        if (chatDetails._id) {
+        setLoading(true);
+        if (chatDetails && chatDetails._id) {
             axios.get(`http://localhost:4000/api/message/${chatDetails._id}`, config)
                 .then((response) => {
                     setMessages(response.data);
-                    socket.emit("join room", chatDetails._id)
-                    setLoading(false)
+                    socket.emit("join room", chatDetails._id);
+                    setLoading(false);
                 })
                 .catch((err) => {
                     console.log(err);
                 });
         }
-    }, [chatDetails, config], selectedChat);
+    }, [chatDetails, config]);
 
-    // function fetchMessages() {
-    //     setLoading(true)
-    //     if (chatDetails._id) {
-    //         console.log(chatDetails)
-    //         axios.get(`http://localhost:4000/api/message/${chatDetails._id}`, config)
-    //             .then((response) => {
-    //                 setMessages(response.data);
-    //                 socket.emit("join room", chatDetails._id)
-    //                 setLoading(false)
-    //                 console.log(response.data)
-    //             })
-    //             .catch((err) => {
-    //                 console.log(err);
-    //             });
-    //     }
-    // }
-
+    console.log(chatDetails)
     return (
         <div className="chat-window">
-            {selectedChat ? (
+            {selectedChat && chatDetails ? (
                 <>
                     <div className="chat-header">
                         <div className="chat-header-info">
-                            <img src={otherUserDetails.profilePhoto} alt="Profile" className="chat-header-image" />
+                            <img src={chatDetails.isGroupChat ? groupIconUrl : otherUserDetails.profilePhoto} alt="Profile" className="chat-header-image" />
                             <div className="chat-header-name">
-                                <h4>{otherUserDetails.name}</h4>
+                                <h4>{chatDetails.isGroupChat ? chatDetails.chatName : otherUserDetails.name}</h4>
                                 <span>
-                                    {isTyping ? <div className="typing-indicator">
-                                        <span>Typing</span>
-                                        <div className="dot"></div>
-                                        <div className="dot"></div>
-                                        <div className="dot"></div>
-                                    </div> : null}
+                                    {isTyping ? (
+                                        <div className="typing-indicator">
+                                            <span>Typing</span>
+                                            <div className="dot"></div>
+                                            <div className="dot"></div>
+                                            <div className="dot"></div>
+                                        </div>
+                                    ) : null}
                                 </span>
                             </div>
                         </div>
-                        <p>chat id:{selectedChat}</p>
+                        {chatDetails.isGroupChat ?
+                            <div style={{display:'flex',gap:6}}>
+                                {chatDetails.users.map((user) => <p>{user.name},</p>)}
+                            </div> : null}
                         <div className='header-menu'>
                             <button className="delete-button" onClick={handleDelete} title="Delete">
                                 <FontAwesomeIcon icon={faTrashAlt} />
@@ -204,25 +188,16 @@ function ChatWindow({ selectedChat, setSelectedChat, setUsersListUpdate}) {
                         </div>
                     </div>
                     <div className="chat-messages">
-                        {loading ? <span className='chat-window-loader'></span> :
-                            <>
-                                {messages.map((message) => (
-                                    <div key={message._id} className={`message ${message.sender._id === userDetails._id ? 'current-user-message' : 'other-user-message'}`}>
-                                        <p>{message.content}</p>
-                                        <div>
-                                            <span className='timestamp'>{new Date(message.createdAt).toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' })}</span>
-                                            {/* <div className={`read-by-ticks ${message.readBy.includes(userDetails._id) ? `blue-ticks` : null}`}>
-                                        <span className='tick-1'><FontAwesomeIcon icon={faCheck} /></span>
-                                        <span className='tick-2'><FontAwesomeIcon icon={faCheck} /></span>
-                                    </div> */}
-                                        </div>
-                                    </div>))}
-                                <div ref={messagesEndRef}></div>
-                            </>
-                        }
-
+                        {messages.map((message) => (
+                            <div key={message._id} className={`message ${message.sender._id === userDetails._id ? 'current-user-message' : 'other-user-message'}`}>
+                                <p>{message.content}</p>
+                                <div>
+                                    <span className='timestamp'>{new Date(message.createdAt).toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' })}</span>
+                                </div>
+                            </div>
+                        ))}
+                        <div ref={messagesEndRef}></div>
                     </div>
-
                     <div className="chat-input">
                         <input
                             type="text"
@@ -231,7 +206,7 @@ function ChatWindow({ selectedChat, setSelectedChat, setUsersListUpdate}) {
                             onChange={handleInputChange}
                             onKeyDown={(e) => {
                                 if (e.code === "Enter") {
-                                    handleSendMessage()
+                                    handleSendMessage();
                                 }
                             }}
                         />
@@ -246,7 +221,6 @@ function ChatWindow({ selectedChat, setSelectedChat, setUsersListUpdate}) {
                     </div>
                 </div>
             )}
-
         </div>
     );
 }
